New functions for default category in project api
project_get_default_category() returns the default category_id for the
specified (or the current) project. It also handles the ALL_PROJECTS
case. If the default category is undefined or invalid, the function
returns 0, except if the project has only 1 category, in which case it
is used as default.

project_set_default_category() updates the project record with a new
default category; for ALL_PROJECT, the value is stored in global config
'default_category_for_moves'

This commit is a prerequisite to resolve #8582