Allow Administrators to login when system is down for maintenance
With this change, Administrators can login while Mantis is down for
maintenance by setting mbadmin=1 in the URL. This special "Maintenance
Bypass" mode is then stored in a cookie, so subsequent navigation is
seamless.

This corrects the previous behavior, where the admin had to append
mbadmin=1 after each page redirection.

To avoid undesired connections by unauthorized users with a valid
session, cookies are cleared when mbadmin is set on the login page and
before the mbadmin cookie is saved, which forces user to reconnect.

While in Maintenance Bypass mode, html_header() displays a banner at the
top of each page to indicate that.

Note: the cookie name 'MANTIS_MBADMIN_COOKIE' is hardcoded in core.php,
and does not take the value of $g_cookie_prefix into consideration. This
can't be avoided as the bypass check takes place before the inclusion of
config_inc.php.

Fixes #8994