Issue #10853: improve sorting of username fields
This involves a complex SQL statement using CASE, COALESCE and CONCAT
functions, which is required to have (at least semi-) proper sorting of
username fields.

The sort key is defined as follows: realname (if $g_show_realname is
defined) if not null, username if not null, and finally
$g_prefix_for_deleted_users + user id (basically replicating the logic
of user_get_name() function).

The foreign key map is now a static variable.