Improve get_font_path() to handle DejaVu fonts
On most Linux systems, the DejaVu font is installed by default whereas
the MS TrueType fonts may not.

The function now checks for 'dejavu' directory below the base font path
prior to checking the 'msttcorefonts'. This gives a better chance to
have a valid fonts directory when $g_system_font_folder is not defined.