Catch errors when executing GraphViz tool
Retrieving output from stderr, and if not empty throw a ServiceException
with details about the error.

Note: we have to rely on a temp file to capture stderr output. This is
due to a limitation on Windows which does not support non-blocking
streams and hangs if there is no error output (stream_set_blocking has
no effect).